clickless/requirements.txt
django>=5.0.2
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
mysqlclient>=2.2.0
numpy>=1.24.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1 